requirements.txt
chainlit==1.0.300
langchain==0.1.16
openai==1.25.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1